backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
langchain==0.1.0
langchain-core==0.1.45
langchain-community==0.0.13
langchain-openai==0.0.2
langchain-groq==0.1.4
langchain-google-genai==1.0.1
langchain-nvidia-ai-endpoints==0.1.0
langgraph==0.0.26
faiss-cpu==1.8.0
ddgs==9.10.0
requests==2.31.0
PyPDF2==3.0.1
pypdf==3.17.0
aiofiles==23.2.1
tavily-python==0.3.1
psycopg2-binary


